# Refactored Perceptionhash for performance