--- a/AUTHORS.md
+++ b/AUTHORS.md
@@ -3,3 +3,4 @@
 - [Dong-hee Na](https://github.com/corona10/) donghee.na92@gmail.com
 - [Gustavo Brunoro](https://github.com/brunoro/) git@hitnail.net
 - [Alex Higashino](https://github.com/TokyoWolFrog/) TokyoWolFrog@mayxyou.com
+- [Evan Oberholster](https://github.com/evanoberholster/) eroberholster@gmail.com