--- a/hashcompute_test.go
+++ b/hashcompute_test.go
@@ -9,6 +9,10 @@ import (
 	"image/jpeg"
 	"os"
 	"testing"
+
+	"github.com/corona10/goimagehash/etcs"
+	"github.com/corona10/goimagehash/transforms"
+	"github.com/nfnt/resize"
 )
 
 func TestHashCompute(t *testing.T) {
@@ -46,6 +50,15 @@ func TestHashCompute(t *testing.T) {
 		{"_examples/sample1.jpg", "_examples/sample4.jpg", PerceptionHash, "PerceptionHash", 30},
 		{"_examples/sample2.jpg", "_examples/sample3.jpg", PerceptionHash, "PerceptionHash", 34},
 		{"_examples/sample2.jpg", "_examples/sample4.jpg", PerceptionHash, "PerceptionHash", 20},
+		{"_examples/sample1.jpg", "_examples/sample1.jpg", PerceptionHashFast, "PerceptionHashFast", 0},
+		{"_examples/sample2.jpg", "_examples/sample2.jpg", PerceptionHashFast, "PerceptionHashFast", 0},
+		{"_examples/sample3.jpg", "_examples/sample3.jpg", PerceptionHashFast, "PerceptionHashFast", 0},
+		{"_examples/sample4.jpg", "_examples/sample4.jpg", PerceptionHashFast, "PerceptionHashFast", 0},
+		{"_examples/sample1.jpg", "_examples/sample2.jpg", PerceptionHashFast, "PerceptionHashFast", 32},
+		{"_examples/sample1.jpg", "_examples/sample3.jpg", PerceptionHashFast, "PerceptionHashFast", 2},
+		{"_examples/sample1.jpg", "_examples/sample4.jpg", PerceptionHashFast, "PerceptionHashFast", 30},
+		{"_examples/sample2.jpg", "_examples/sample3.jpg", PerceptionHashFast, "PerceptionHashFast", 34},
+		{"_examples/sample2.jpg", "_examples/sample4.jpg", PerceptionHashFast, "PerceptionHashFast", 20},
 	} {
 		file1, err := os.Open(tt.img1)
 		if err != nil {
@@ -94,6 +107,78 @@ func TestHashCompute(t *testing.T) {
 
 		if dis1 != tt.distance {
 			t.Errorf("%s: Distance between %v and %v is expected %v but got %v", tt.name, tt.img1, tt.img2, tt.distance, dis1)
+		}
+	}
+}
+
+func TestHashComputeFast(t *testing.T) {
+	for _, tt := range []struct {
+		img1 string
+	}{
+		{"_examples/sample1.jpg"},
+		{"_examples/sample2.jpg"},
+		{"_examples/sample3.jpg"},
+		{"_examples/sample4.jpg"},
+	} {
+		file1, err := os.Open(tt.img1)
+		if err != nil {
+			t.Errorf("%s", err)
+		}
+		defer file1.Close()
+
+		img1, err := jpeg.Decode(file1)
+		if err != nil {
+			t.Errorf("%s", err)
+		}
+
+		resized := resize.Resize(64, 64, img1, resize.Bilinear)
+
+		hash1 := NewImageHash(0, PHash)
+		pixels := transforms.Rgb2Gray(resized)
+		dct := transforms.DCT2D(pixels, 64, 64)
+		flattens := transforms.FlattenPixels(dct, 8, 8)
+		median1 := etcs.MedianOfPixels(flattens)
+
+		for idx, p := range flattens {
+			if p > median1 {
+				hash1.leftShiftSet(len(flattens) - idx - 1)
+			}
+		}
+
+		p := make([]float64, 4096)
+		pixelsFast := &p
+
+		hash2 := NewImageHash(0, PHash)
+		transforms.Rgb2GrayFast(resized, pixelsFast)
+		transforms.DCT2DFast64(pixelsFast)
+		flattensFast := transforms.FlattenPixelsFast64(*pixelsFast, 8, 8)
+		median2 := etcs.MedianOfPixels(flattensFast)
+
+		for idx, p := range flattensFast {
+			if p > median2 {
+				hash2.leftShiftSet(len(flattensFast) - idx - 1)
+			}
+		}
+
+		if median1 != median2 {
+			t.Errorf("Medians should be identical %v vs %v", median1, median2)
+		}
+
+		dis1, err := hash1.Distance(hash2)
+		if err != nil {
+			t.Errorf("%s", err)
+		}
+
+		dis2, err := hash2.Distance(hash1)
+		if err != nil {
+			t.Errorf("%s", err)
+		}
+
+		if dis1 != dis2 {
+			t.Errorf("Distance should be identical %v vs %v", dis1, dis2)
+		}
+		if dis2 != 0 {
+			t.Errorf("Distance should be identical %v vs %v, with distance of %d", hash1.ToString(), hash2.ToString(), dis1)
 		}
 	}
 }
@@ -324,6 +409,24 @@ func BenchmarkPerceptionHash(b *testing.B) {
 	}
 	for i := 0; i < b.N; i++ {
 		_, err := ExtPerceptionHash(img1, 8, 8)
+		if err != nil {
+			b.Errorf("%s", err)
+		}
+	}
+}
+
+func BenchmarkPerceptionHashFast(b *testing.B) {
+	file1, err := os.Open("_examples/sample3.jpg")
+	if err != nil {
+		b.Errorf("%s", err)
+	}
+	defer file1.Close()
+	img1, err := jpeg.Decode(file1)
+	if err != nil {
+		b.Errorf("%s", err)
+	}
+	for i := 0; i < b.N; i++ {
+		_, err := PerceptionHashFast(img1)
 		if err != nil {
 			b.Errorf("%s", err)
 		}