--- a/transforms/dct.go
+++ b/transforms/dct.go
@@ -73,3 +73,49 @@ func DCT2D(input [][]float64, w int, h int) [][]float64 {
 	wg.Wait()
 	return output
 }
+
+// DCT2DFast64 function returns a result of DCT2D by using the seperable property.
+// Fast uses static DCT tables for improved performance.
+func DCT2DFast64(input *[]float64) {
+	if len(*input) != 4096 {
+		panic("incorrect input size")
+	}
+
+	for i := 0; i < 64; i++ { // height
+		DCT1DFast64((*input)[i*64 : (i*64)+64])
+		//forwardTransformFast((*input)[i*64:(i*64)+64], temp[:], 64)
+	}
+
+	for i := 0; i < 64; i++ { // width
+		row := [64]float64{}
+		for j := 0; j < 64; j++ {
+			row[j] = (*input)[i+((j)*64)]
+		}
+		DCT1DFast64(row[:])
+		for j := 0; j < len(row); j++ {
+			(*input)[i+(j*64)] = row[j]
+		}
+	}
+}
+
+func forwardTransformFast(input, temp []float64, Len int) {
+	if Len == 1 {
+		return
+	}
+
+	halfLen := Len / 2
+	t := dctTables[halfLen>>1]
+	for i := 0; i < halfLen; i++ {
+		x, y := input[i], input[Len-1-i]
+		temp[i] = x + y
+		temp[i+halfLen] = (x - y) / t[i]
+	}
+	forwardTransformFast(temp, input, halfLen)
+	forwardTransformFast(temp[halfLen:], input, halfLen)
+	for i := 0; i < halfLen-1; i++ {
+		input[i*2+0] = temp[i]
+		input[i*2+1] = temp[i+halfLen] + temp[i+halfLen+1]
+	}
+
+	input[Len-2], input[Len-1] = temp[halfLen-1], temp[Len-1]
+}