Refactored PerceptionHash for performance. If you want me to change the name to replace existing function that would be acceptable.

Features:
- Significantly reduces allocations
- Uses static DCT1D tables
- Uses sync pool to reduce pixel data allocations
- Uses image.Image optimizations for obtaining pixels for improved inlining of functions
- Backward compatible with PerceptionHash, they are interchangeable. Confirmed with passing tests.
- Performance without image resize as below.
```go
name                      time/op
PerceptionHash/Regular-12         358µs ±26%
PerceptionHash/Fast-12           84.8µs ± 7%
PerceptionHash/Parallel-12       82.5µs ± 3%
PerceptionHash/Fast-Parallel-12  11.5µs ± 1%

name                      alloc/op
PerceptionHash/Regular-12         193kB ± 0%
PerceptionHash/Fast-12            17.0B ± 0%
PerceptionHash/Parallel-12        193kB ± 0%
PerceptionHash/Fast-Parallel-12    198B ± 0%

name                      allocs/op
PerceptionHash/Regular-12         4.68k ± 0%
PerceptionHash/Fast-12             0.00     
PerceptionHash/Parallel-12        4.68k ± 0%
PerceptionHash/Fast-Parallel-12    0.00 
```