@evanoberholster 2022-05-12
Comparison between PerceptionHash and PerceptionHashFast:
```go
name               old time/op    new time/op    delta
PerceptionHash-12     945µs ± 9%     689µs ± 2%  -27.08%  (p=0.000 n=170+177)

name               old alloc/op   new alloc/op   delta
PerceptionHash-12     471kB ± 0%     298kB ± 0%  -36.80%  (p=0.000 n=174+198)

name               old allocs/op  new allocs/op  delta
PerceptionHash-12     4.74k ± 0%     0.07k ± 0%  -98.61%  (p=0.000 n=200+200)
```