@corona10 2022-05-12
Please add your name to the following file.
https://github.com/corona10/goimagehash/blob/master/AUTHORS.md