@corona10 2022-05-12
> Significantly reduces allocations
> Uses static DCT1D tables
> Uses sync pool to reduce pixel data allocations
> Uses image.Image optimizations for obtaining pixels for improved inlining of functions
> Backward compatible with PerceptionHash, they are interchangeable. Confirmed with passing tests.
> Performance without image resize as below.

Thanks, I love all of your changes :)