@evanoberholster 2022-05-15
PerceptionHashFast DCT static tables only support a size of 64x64. We would need to include further static tables to support other image sizes, (eg 256x256)