@corona10 2022-05-15
@evanoberholster 

> PerceptionHashFast DCT static tables only support a size of 64x64.

Ah okay, let's replace PerceptionHash only.